+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Patches have coordinates for visualization; calculated if not provided (#65)
 - Daily patch visits for pollen and nectar are counted, adds respective observers (#66)
 - Stats for foraging rounds are recorded, adds respective table observer (#67)
+- Move `CustomParams` from beecs-cli to the core module (#68)
 
 ### Bugfixes