+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [[unpublished]](https://github.com/mlange-42/beecs/compare/v0.2.0...main)
+## [[v0.3.0]](https://github.com/mlange-42/beecs/compare/v0.2.0...v0.3.0)
 
 ### Breaking changes